requirements.txt
tensorflow>=2.9.0
torch>=1.12.0
torchvision>=0.13.0
transformers>=4.25.1
diffusers>=0.14.0
accelerate>=0.16.0
pillow>=9.0.0
numpy>=1.21.0
tqdm>=4.64.0
gradio>=3.6.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
huggingface-hub>=0.12.0 